@@ -6,15 +6,18 @@ pub enum SyncRequestFile {
     Add {
         path: String,
         content: Vec<u8>,
+        src_cr: Option<usize>,
     },
     Update {
         path: String,
         content: Vec<u8>,
         version: i32,
+        src_cr: Option<usize>,
     },
     Delete {
         path: String,
         version: i32,
+        src_cr: Option<usize>,
     },
 }
 
@@ -32,7 +35,7 @@ pub struct SyncResponse {
     pub latest_ftd: String,
 }
 
-#[derive(serde::Serialize, serde::Deserialize, std::fmt::Debug, PartialEq)]
+#[derive(serde::Serialize, serde::Deserialize, std::fmt::Debug, PartialEq, Eq)]
 pub enum SyncStatus {
     RegularConflict,
     NoConflict,
@@ -122,20 +125,25 @@ pub async fn sync2(
     }
 }
 
-pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncResponse> {
-    // TODO: Need to call at once only
-    let config = fpm::Config::read(None, false).await?;
-    let mut server_history = config.get_history().await?;
-    let server_latest =
-        fpm::history::FileHistory::get_non_deleted_latest_file_edits(server_history.as_slice())?;
+pub(crate) async fn do_sync(
+    config: &fpm::Config,
+    files: &[SyncRequestFile],
+) -> fpm::Result<std::collections::HashMap<String, SyncResponseFile>> {
+    let mut remote_history = config.get_history().await?;
+    let remote_manifest =
+        fpm::history::FileHistory::get_remote_manifest(remote_history.as_slice(), false)?;
     let mut to_be_in_history: std::collections::BTreeMap<String, fpm::history::FileEditTemp> =
         Default::default();
     let mut synced_files = std::collections::HashMap::new();
-    for file in request.files {
+    for file in files {
         // TODO: get all data like message, author, src-cr from request
-        match &file {
-            SyncRequestFile::Add { path, content } => {
-                if server_latest.contains_key(path) {
+        match file {
+            SyncRequestFile::Add {
+                path,
+                content,
+                src_cr,
+            } => {
+                if remote_manifest.contains_key(path) {
                     // add-add-conflict
                     synced_files.insert(
                         path.to_string(),
@@ -153,7 +161,7 @@ pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncRespons
                     fpm::history::FileEditTemp {
                         message: None,
                         author: None,
-                        src_cr: None,
+                        src_cr: *src_cr,
                         operation: fpm::history::FileOperation::Added,
                     },
                 );
@@ -162,8 +170,9 @@ pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncRespons
                 path,
                 content,
                 version,
+                src_cr,
             } => {
-                if let Some(file_edit) = server_latest.get(path) {
+                if let Some(file_edit) = remote_manifest.get(path) {
                     if file_edit.version.eq(version) {
                         fpm::utils::update(&config.root.join(path), content).await?;
                         // TODO: get all data like message, author, src-cr from request
@@ -172,7 +181,7 @@ pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncRespons
                             fpm::history::FileEditTemp {
                                 message: None,
                                 author: None,
-                                src_cr: None,
+                                src_cr: *src_cr,
                                 operation: fpm::history::FileOperation::Updated,
                             },
                         );
@@ -211,7 +220,7 @@ pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncRespons
                                     fpm::history::FileEditTemp {
                                         message: None,
                                         author: None,
-                                        src_cr: None,
+                                        src_cr: *src_cr,
                                         operation: fpm::history::FileOperation::Updated,
                                     },
                                 );
@@ -252,8 +261,12 @@ pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncRespons
                     continue;
                 };
             }
-            SyncRequestFile::Delete { path, version } => {
-                let file_edit = if let Some(file_edit) = server_latest.get(path) {
+            SyncRequestFile::Delete {
+                path,
+                version,
+                src_cr,
+            } => {
+                let file_edit = if let Some(file_edit) = remote_manifest.get(path) {
                     file_edit
                 } else {
                     // ALready deleted in server, do nothing
@@ -282,7 +295,7 @@ pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncRespons
                         fpm::history::FileEditTemp {
                             message: None,
                             author: None,
-                            src_cr: None,
+                            src_cr: *src_cr,
                             operation: fpm::history::FileOperation::Deleted,
                         },
                     );
@@ -291,18 +304,25 @@ pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncRespons
         }
     }
 
-    fpm::history::insert_into_history(&config.root, &to_be_in_history, &mut server_history).await?;
+    fpm::history::insert_into_history(&config.root, &to_be_in_history, &mut remote_history).await?;
+    Ok(synced_files)
+}
 
-    let server_latest =
-        fpm::history::FileHistory::get_latest_file_edits(server_history.as_slice())?;
+pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncResponse> {
+    // TODO: Need to call at once only
+    let config = fpm::Config::read(None, false).await?;
+    let mut synced_files = do_sync(&config, request.files.as_slice()).await?;
+    let remote_history = config.get_history().await?;
+    let remote_manifest =
+        fpm::history::FileHistory::get_remote_manifest(remote_history.as_slice(), true)?;
 
-    let client_history = fpm::history::FileHistory::from_ftd(request.history.as_str())?;
+    let clone_history = fpm::history::FileHistory::from_ftd(request.history.as_str())?;
     let client_latest =
-        fpm::history::FileHistory::get_latest_file_edits(client_history.as_slice())?;
+        fpm::history::FileHistory::get_remote_manifest(clone_history.as_slice(), true)?;
 
-    client_current_files(&config, &server_latest, &client_latest, &mut synced_files).await?;
+    client_current_files(&config, &remote_manifest, &client_latest, &mut synced_files).await?;
 
-    let history_files = client_history_files(&config, &server_latest, &client_latest).await?;
+    let history_files = clone_history_files(&config, &remote_manifest, &client_latest).await?;
 
     Ok(SyncResponse {
         files: synced_files.into_values().collect_vec(),
@@ -311,13 +331,14 @@ pub(crate) async fn sync_worker(request: SyncRequest) -> fpm::Result<SyncRespons
     })
 }
 
-async fn client_history_files(
+async fn clone_history_files(
     config: &fpm::Config,
-    server_latest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
+    remote_manifest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
     client_latest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
 ) -> fpm::Result<Vec<File>> {
-    let diff = snapshot_diff(server_latest, client_latest);
-    let history = ignore::WalkBuilder::new(config.server_history_dir())
+    let diff = snapshot_diff(remote_manifest, client_latest);
+    let history = ignore::WalkBuilder::new(config.remote_history_dir())
+        .hidden(false)
         .build()
         .into_iter()
         .flatten()
@@ -325,7 +346,7 @@ async fn client_history_files(
             x.into_path()
                 .to_str()
                 .unwrap()
-                .trim_start_matches(config.server_history_dir().as_str())
+                .trim_start_matches(config.remote_history_dir().as_str())
                 .trim_matches('/')
                 .to_string()
         })
@@ -339,7 +360,7 @@ async fn client_history_files(
             .filter(|x| client_file_edit.map(|c| x.0.gt(&c.version)).unwrap_or(true))
             .collect_vec();
         for (_, path) in history_paths {
-            let content = tokio::fs::read(config.server_history_dir().join(&path)).await?;
+            let content = tokio::fs::read(config.remote_history_dir().join(&path)).await?;
             dot_history.push(File { path, content });
         }
     }
@@ -369,11 +390,11 @@ fn get_all_versions(path: &str, history: &[String]) -> fpm::Result<Vec<(i32, Str
 
 async fn client_current_files(
     config: &fpm::Config,
-    server_latest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
+    remote_manifest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
     client_latest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
     synced_files: &mut std::collections::HashMap<String, SyncResponseFile>,
 ) -> fpm::Result<()> {
-    let diff = snapshot_diff(server_latest, client_latest);
+    let diff = snapshot_diff(remote_manifest, client_latest);
     for (path, operation) in diff.iter() {
         if synced_files.contains_key(path) {
             continue;
@@ -403,7 +424,7 @@ async fn client_current_files(
     // Deleted files
     let diff = client_latest
         .iter()
-        .filter(|(path, _)| !server_latest.contains_key(path.as_str()));
+        .filter(|(path, _)| !remote_manifest.contains_key(path.as_str()));
 
     // TODO: If already in synced files need to handle that case
     for (path, _) in diff {
@@ -422,12 +443,12 @@ async fn client_current_files(
 }
 
 fn snapshot_diff(
-    server_latest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
+    remote_manifest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
     client_latest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
 ) -> std::collections::BTreeMap<String, fpm::history::FileOperation> {
     let mut diff: std::collections::BTreeMap<String, fpm::history::FileOperation> =
         Default::default();
-    for (snapshot_path, file_edit) in server_latest {
+    for (snapshot_path, file_edit) in remote_manifest {
         match client_latest.get(snapshot_path) {
             Some(client_file_edit) if client_file_edit.version.lt(&file_edit.version) => {
                 diff.insert(snapshot_path.to_string(), file_edit.operation.clone());
