@@ -1,22 +1,14 @@
 pub(crate) async fn view_source(req: actix_web::HttpRequest) -> actix_web::HttpResponse {
     // TODO: Need to remove unwrap
     let path = {
-        let mut path: std::path::PathBuf = req.match_info().query("path").parse().unwrap();
-        if path.eq(&std::path::PathBuf::new().join("")) {
+        let mut path: camino::Utf8PathBuf = req.match_info().query("path").parse().unwrap();
+        if path.eq(&camino::Utf8PathBuf::new().join("")) {
             path = path.join("/");
         }
         path
     };
 
-    let path = match path.to_str() {
-        Some(s) => s,
-        None => {
-            println!("handle_ftd: Not able to convert path");
-            return actix_web::HttpResponse::InternalServerError().body("".as_bytes());
-        }
-    };
-
-    match handle_view_source(path).await {
+    match handle_view_source(path.as_str()).await {
         Ok(body) => actix_web::HttpResponse::Ok().body(body),
         Err(e) => {
             println!("new_path: {}, Error: {:?}", path, e);
@@ -25,29 +17,41 @@ pub(crate) async fn view_source(req: actix_web::HttpRequest) -> actix_web::HttpR
     }
 }
 
+pub(crate) async fn _view_source_h(
+    _req: hyper::Request<hyper::Body>,
+) -> hyper::Response<hyper::Body> {
+    // TODO: Need to remove unwrap
+    let path: String = {
+        // let mut path: camino::Utf8PathBuf = req.match_info().query("path").parse().unwrap();
+        // if path.eq(&camino::Utf8PathBuf::new().join("")) {
+        //     path = path.join("/");
+        // }
+        // path
+        "TODO".to_string()
+    };
+
+    match handle_view_source(path.as_str()).await {
+        Ok(body) => hyper::Response::new(body.into()),
+        Err(e) => {
+            println!("new_path: {}, Error: {:?}", path, e);
+            fpm::apis::response::_error_h(e.to_string(), hyper::StatusCode::INTERNAL_SERVER_ERROR)
+        }
+    }
+}
+
 async fn handle_view_source(path: &str) -> fpm::Result<Vec<u8>> {
     let mut config = fpm::Config::read(None, false).await?;
     let file_name = config.get_file_path_and_resolve(path).await?;
     let file = config.get_file_and_package_by_id(path).await?;
-    let editor_ftd = if config.root.join("e.ftd").exists() {
-        tokio::fs::read_to_string(config.root.join("e.ftd")).await?
-    } else {
-        fpm::editor_ftd().to_string()
-    };
 
     match file {
         fpm::File::Ftd(_) | fpm::File::Markdown(_) | fpm::File::Code(_) => {
             let snapshots = fpm::snapshot::get_latest_snapshots(&config.root).await?;
-            let mut editor_content = format!(
-                "{}\n\n-- source:\n$processor$: fetch-file\npath:{}\n\n-- path: {}\n\n",
-                editor_ftd, file_name, file_name,
-            );
-            if let Ok(Some(diff)) = get_diff(&file, &snapshots).await {
-                editor_content = format!("{}\n\n\n-- diff:\n\n{}", editor_content, diff);
-            }
+            let diff = get_diff(&file, &snapshots).await;
+            let editor_ftd = fpm::package_info_editor(&config, file_name.as_str(), diff)?;
             let main_document = fpm::Document {
                 id: "editor.ftd".to_string(),
-                content: editor_content,
+                content: editor_ftd,
                 parent_path: config.root.as_str().to_string(),
                 package_name: config.package.name.clone(),
             };
