@@ -7,6 +7,9 @@ pub async fn build2(
     tokio::fs::create_dir_all(config.build_dir()).await?;
     let documents = get_documents_for_current_package(config).await?;
 
+    // No need to build static files when file is passed during fpm build (no-static behaviour)
+    let no_static: bool = file.is_some();
+
     for main in documents.values() {
         if file.is_some() && file != Some(main.get_id().as_str()) {
             continue;
@@ -19,9 +22,6 @@ pub async fn build2(
             config.package.name.as_str(),
             main.get_id()
         );
-
-        // No need to build static files when file is passed during fpm build (no-static behaviour)
-        let no_static: bool = file.is_some();
 
         match main {
             fpm::File::Ftd(doc) => {
@@ -99,14 +99,16 @@ pub async fn build2(
             start,
         );
     }
+
+    if !no_static {
+        config.download_fonts().await?;
+    }
     Ok(())
 }
 
 async fn get_documents_for_current_package(
     config: &mut fpm::Config,
 ) -> fpm::Result<std::collections::BTreeMap<String, fpm::File>> {
-    use std::iter::FromIterator;
-
     let mut documents = std::collections::BTreeMap::from_iter(
         config
             .get_files(&config.package)
@@ -269,13 +271,15 @@ async fn process_markdown(
         doc: &fpm::Document,
     ) -> fpm::Result<Option<fpm::Document>> {
         let doc_id = if doc.id == "README.md"
-            && !(std::path::Path::new(format!(".{}index.ftd", std::path::MAIN_SEPARATOR).as_str())
+            && !(camino::Utf8Path::new(format!(".{}index.ftd", std::path::MAIN_SEPARATOR).as_str())
                 .exists()
-                || std::path::Path::new(format!(".{}index.md", std::path::MAIN_SEPARATOR).as_str())
-                    .exists())
+                || camino::Utf8Path::new(
+                    format!(".{}index.md", std::path::MAIN_SEPARATOR).as_str(),
+                )
+                .exists())
         {
             "index.md".to_string()
-        } else if !std::path::Path::new(
+        } else if !camino::Utf8Path::new(
             format!(
                 ".{}{}",
                 std::path::MAIN_SEPARATOR,
