@@ -5,13 +5,13 @@ use std::iter::FromIterator;
 
 #[derive(Debug, Clone)]
 pub struct Config {
-    pub package: fpm::Package,
+    pub package: Package,
     pub root: camino::Utf8PathBuf,
     pub packages_root: camino::Utf8PathBuf,
     pub original_directory: camino::Utf8PathBuf,
     pub extra_data: serde_json::Map<String, serde_json::Value>,
     pub current_document: Option<String>,
-    pub all_packages: std::cell::RefCell<std::collections::BTreeMap<String, fpm::Package>>,
+    pub all_packages: std::cell::RefCell<std::collections::BTreeMap<String, Package>>,
     pub downloaded_assets: std::collections::BTreeMap<String, String>,
     pub global_ids: std::collections::HashMap<String, String>,
 }
@@ -36,23 +36,61 @@ impl Config {
     }
 
     pub fn cr_path(&self, cr_number: usize) -> camino::Utf8PathBuf {
-        self.root.join("-/").join(cr_number.to_string())
+        self.root.join(fpm::cr::cr_path(cr_number))
+    }
+
+    pub fn path_without_root(&self, path: &camino::Utf8PathBuf) -> fpm::Result<String> {
+        Ok(path.strip_prefix(&self.root)?.to_string())
+    }
+
+    pub fn cr_deleted_file_path(&self, cr_number: usize) -> camino::Utf8PathBuf {
+        self.cr_path(cr_number).join("-/deleted.ftd")
+    }
+
+    pub fn track_path(&self, path: &camino::Utf8PathBuf) -> camino::Utf8PathBuf {
+        let path_without_root = self
+            .path_without_root(path)
+            .unwrap_or_else(|_| path.to_string());
+        let track_path = format!("{}.track", path_without_root);
+        self.track_dir().join(track_path)
+    }
+
+    pub fn cr_track_dir(&self, cr_number: usize) -> camino::Utf8PathBuf {
+        self.track_dir().join(fpm::cr::cr_path(cr_number))
+    }
+
+    pub fn cr_track_path(
+        &self,
+        path: &camino::Utf8PathBuf,
+        cr_number: usize,
+    ) -> camino::Utf8PathBuf {
+        let path_without_root = self
+            .cr_path(cr_number)
+            .join(path)
+            .to_string()
+            .replace(self.root.to_string().as_str(), "");
+        let track_path = format!("{}.track", path_without_root);
+        self.track_dir().join(track_path)
     }
 
     pub fn cr_about_path(&self, cr_number: usize) -> camino::Utf8PathBuf {
         self.cr_path(cr_number).join("-/about.ftd")
     }
 
+    pub fn cr_meta_path(&self, cr_number: usize) -> camino::Utf8PathBuf {
+        self.cr_path(cr_number).join("-/meta.ftd")
+    }
+
     pub fn remote_dir(&self) -> camino::Utf8PathBuf {
         self.root.join(".remote-state")
     }
 
-    pub fn server_history_dir(&self) -> camino::Utf8PathBuf {
+    pub fn remote_history_dir(&self) -> camino::Utf8PathBuf {
         self.remote_dir().join("history")
     }
 
     /// location that stores lowest available cr number
-    pub fn server_cr(&self) -> camino::Utf8PathBuf {
+    pub fn remote_cr(&self) -> camino::Utf8PathBuf {
         self.remote_dir().join("cr")
     }
 
@@ -62,7 +100,24 @@ impl Config {
 
     pub(crate) fn history_path(&self, id: &str, version: i32) -> camino::Utf8PathBuf {
         let id_with_timestamp_extension = fpm::utils::snapshot_id(id, &(version as u128));
-        self.server_history_dir().join(id_with_timestamp_extension)
+        self.remote_history_dir().join(id_with_timestamp_extension)
+    }
+
+    /// document_name_with_default("index.ftd") -> /
+    /// document_name_with_default("foo/index.ftd") -> /foo/
+    /// document_name_with_default("foo/abc") -> /foo/abc/
+    /// document_name_with_default("/foo/abc.ftd") -> /foo/abc/
+    pub(crate) fn document_name_with_default(&self, document_path: &str) -> String {
+        let name = self
+            .doc_id()
+            .unwrap_or_else(|| document_path.to_string())
+            .trim_matches('/')
+            .to_string();
+        if name.is_empty() {
+            "/".to_string()
+        } else {
+            format!("/{}/", name)
+        }
     }
 
     /// history of a fpm package is stored in `.history` folder.
@@ -92,7 +147,7 @@ impl Config {
         self.fpm_dir().join("conflicted")
     }
 
-    /// every package's `.history` contains a file `.latest.ftd`. It looks a bit linke this:
+    /// every package's `.history` contains a file `.latest.ftd`. It looks a bit link this:
     ///
     /// ```ftd
     /// -- import: fpm
@@ -138,11 +193,9 @@ impl Config {
                 .parent()
                 .expect("Expect fpm_path parent. Panic!")
                 .to_owned()),
-            _ => {
-                return Err(fpm::Error::UsageError {
-                    message: format!("Unable to find `fpm_path` of the package {}", o.name),
-                })
-            }
+            _ => Err(fpm::Error::UsageError {
+                message: format!("Unable to find `fpm_path` of the package {}", o.name),
+            }),
         }
     }
 
@@ -201,6 +254,36 @@ impl Config {
             false => format!("<style>{}</style>", generated_style),
             _ => "".to_string(),
         };
+    }
+
+    pub(crate) async fn download_fonts(&self) -> fpm::Result<()> {
+        use itertools::Itertools;
+
+        let mut fonts = vec![];
+        for dep in self
+            .package
+            .get_flattened_dependencies()
+            .into_iter()
+            .unique_by(|dep| dep.package.name.clone())
+        {
+            fonts.extend(dep.package.fonts);
+        }
+
+        for package in self.all_packages.borrow().values() {
+            fonts.extend(package.fonts.clone());
+        }
+
+        for font in fonts.iter() {
+            if let Some(url) = font.get_url() {
+                let start = std::time::Instant::now();
+                print!("Processing {} ... ", url);
+                let content = self.get_file_and_resolve(url.as_str()).await?.1;
+                fpm::utils::update(&self.build_dir().join(&url), content.as_slice()).await?;
+                fpm::utils::print_end(format!("Processed {}", url).as_str(), start);
+            }
+        }
+
+        Ok(())
     }
 
     /// `attach_data_string()` sets the value of extra data in fpm::Config,
@@ -451,14 +534,48 @@ impl Config {
 
     pub async fn get_file_by_id(&self, id: &str, package: &fpm::Package) -> fpm::Result<fpm::File> {
         let file_name = fpm::Config::get_file_name(&self.root, id)?;
-        return self
-            .get_files(package)
+        self.get_files(package)
             .await?
             .into_iter()
             .find(|v| v.get_id().eq(file_name.as_str()))
             .ok_or_else(|| fpm::Error::UsageError {
                 message: format!("No such file found: {}", id),
-            });
+            })
+    }
+
+    pub(crate) async fn get_file_and_package_by_cr_id(
+        &mut self,
+        id: &str,
+        cr_number: usize,
+    ) -> fpm::Result<fpm::File> {
+        let file_name = self.get_cr_file_and_resolve(id, cr_number).await?.0;
+        let id_without_cr_prefix = fpm::cr::get_id_from_cr_id(id, cr_number)?;
+        let package = self
+            .find_package_by_id(id_without_cr_prefix.as_str())
+            .await?
+            .1;
+
+        let mut file = fpm::get_file(
+            package.name.to_string(),
+            &self.root.join(file_name),
+            &self.get_root_for_package(&package),
+        )
+        .await?;
+
+        if id_without_cr_prefix.contains("-/") && !id_without_cr_prefix.contains("-/about") {
+            let url = id_without_cr_prefix
+                .trim_end_matches("/index.html")
+                .trim_matches('/');
+            let extension = if matches!(file, fpm::File::Markdown(_)) {
+                "/index.md".to_string()
+            } else if matches!(file, fpm::File::Ftd(_)) {
+                "/index.ftd".to_string()
+            } else {
+                "".to_string()
+            };
+            file.set_id(format!("{}{}", url, extension).as_str());
+        }
+        Ok(file)
     }
 
     pub async fn get_file_and_package_by_id(&mut self, id: &str) -> fpm::Result<fpm::File> {
@@ -524,6 +641,10 @@ impl Config {
     }
 
     pub(crate) async fn get_file_path_and_resolve(&self, id: &str) -> fpm::Result<String> {
+        Ok(self.get_file_and_resolve(id).await?.0)
+    }
+
+    pub(crate) async fn get_file_and_resolve(&self, id: &str) -> fpm::Result<(String, Vec<u8>)> {
         let (package_name, package) = self.find_package_by_id(id).await?;
         let package = self.resolve_package(&package).await?;
         self.add_package(&package);
@@ -551,11 +672,71 @@ impl Config {
             id
         };
 
-        Ok(format!(
-            "{}{}",
-            add_packages,
-            package.resolve_by_id(id, None).await?.0
-        ))
+        let (file_name, content) = package.resolve_by_id(id, None).await?;
+
+        Ok((format!("{}{}", add_packages, file_name), content))
+    }
+
+    pub(crate) async fn get_cr_file_and_resolve(
+        &self,
+        cr_id: &str,
+        cr_number: usize,
+    ) -> fpm::Result<(String, Vec<u8>)> {
+        let id_without_cr_prefix = fpm::cr::get_id_from_cr_id(cr_id, cr_number)?;
+        let (package_name, package) = self
+            .find_package_by_id(id_without_cr_prefix.as_str())
+            .await?;
+        let package = self.resolve_package(&package).await?;
+        self.add_package(&package);
+        let mut new_id = id_without_cr_prefix.to_string();
+        let mut add_packages = "".to_string();
+        if let Some(id) = new_id.strip_prefix("-/") {
+            // Check if the id is alias for index.ftd. eg: `/-/bar/`
+            if id.starts_with(&package_name) || !package.name.eq(self.package.name.as_str()) {
+                new_id = id.to_string();
+            }
+            if !package.name.eq(self.package.name.as_str()) {
+                add_packages = format!(".packages/{}/", package.name);
+            }
+        }
+        let id = {
+            let mut id = match new_id.split_once("-/") {
+                Some((p1, p2))
+                    if !(package_name.eq(self.package.name.as_str())
+                        && fpm::utils::ids_matches(p2, "about")) =>
+                // full id in case of about page as it's a special page
+                {
+                    p1.to_string()
+                }
+                _ => new_id,
+            }
+            .trim()
+            .trim_start_matches(package_name.as_str())
+            .to_string();
+            if id.is_empty() {
+                id = "/".to_string();
+            }
+            id
+        };
+
+        if package.name.eq(self.package.name.as_str()) {
+            let file_info_map = fpm::cr::cr_clone_file_info(self, cr_number).await?;
+            let file_info = fpm::package_doc::file_id_to_names(id.as_str())
+                .into_iter()
+                .find_map(|id| file_info_map.get(&id))
+                .ok_or_else(|| fpm::Error::UsageError {
+                    message: format!("{} is not found", cr_id),
+                })?;
+
+            return Ok((
+                format!("{}{}", add_packages, file_info.path),
+                file_info.content.to_owned(),
+            ));
+        }
+
+        let (file_name, content) = package.resolve_by_id(id.as_str(), None).await?;
+
+        Ok((format!("{}{}", add_packages, file_name), content))
     }
 
     /// Return (package name or alias, package)
@@ -755,14 +936,13 @@ impl Config {
 
     pub(crate) async fn get_assets(
         &self,
-        base_url: &str,
     ) -> fpm::Result<std::collections::HashMap<String, String>> {
         use itertools::Itertools;
 
         let mut asset_documents = std::collections::HashMap::new();
         asset_documents.insert(
             self.package.name.clone(),
-            self.package.get_assets_doc(self, base_url).await?,
+            self.package.get_assets_doc().await?,
         );
 
         let dependencies = if let Some(package) = self.package.translation_of.as_ref() {
@@ -788,7 +968,7 @@ impl Config {
         for dep in &dependencies {
             asset_documents.insert(
                 dep.package.name.clone(),
-                dep.package.get_assets_doc(self, base_url).await?,
+                dep.package.get_assets_doc().await?,
             );
         }
         Ok(asset_documents)
@@ -894,14 +1074,13 @@ impl Config {
                 }
             };
 
-            if package.name != fpm::PACKAGE_INFO_INTERFACE
-                && !deps.iter().any(|dep| {
-                    dep.implements
-                        .contains(&fpm::PACKAGE_INFO_INTERFACE.to_string())
-                })
+            if package.name != fpm::FPM_UI_INTERFACE
+                && !deps
+                    .iter()
+                    .any(|dep| dep.implements.contains(&fpm::FPM_UI_INTERFACE.to_string()))
             {
                 deps.push(fpm::Dependency {
-                    package: fpm::Package::new(fpm::PACKAGE_INFO_INTERFACE),
+                    package: fpm::Package::new(fpm::FPM_UI_INTERFACE),
                     version: None,
                     notes: None,
                     alias: None,
@@ -957,7 +1136,7 @@ impl Config {
             global_ids: Default::default(),
         };
 
-        let asset_documents = config.get_assets("/").await?;
+        let asset_documents = config.get_assets().await?;
 
         config.package.sitemap = {
             let sitemap = match package.translation_of.as_ref() {
@@ -967,7 +1146,7 @@ impl Config {
             .sitemap_temp
             .as_ref();
 
-            let sitemap = match sitemap {
+            match sitemap {
                 Some(sitemap_temp) => {
                     let mut s = fpm::sitemap::Sitemap::parse(
                         sitemap_temp.body.as_str(),
@@ -983,8 +1162,7 @@ impl Config {
                     Some(s)
                 }
                 None => None,
-            };
-            sitemap
+            }
         };
 
         config.add_package(&package);
@@ -1045,7 +1223,7 @@ impl Config {
             return fpm::usage_error("Can be used by remote only".to_string());
         }
         let value = fpm::cache::update(
-            self.server_cr().to_string().as_str(),
+            self.remote_cr().to_string().as_str(),
             number_of_crs_to_reserve,
         )
         .await? as i32;
@@ -1053,6 +1231,53 @@ impl Config {
         Ok(Vec::from_iter(
             (value - (number_of_crs_to_reserve as i32))..value,
         ))
+    }
+
+    pub(crate) async fn can_read(
+        &self,
+        req: &actix_web::HttpRequest,
+        document_path: &str,
+    ) -> fpm::Result<bool> {
+        use itertools::Itertools;
+        let document_name = self.document_name_with_default(document_path);
+        let access_identities =
+            fpm::user_group::access_identities(self, req, &document_name, true).await?;
+        if let Some(sitemap) = &self.package.sitemap {
+            // TODO: This can be buggy in case of: if groups are used directly in sitemap are foreign groups
+            let document_readers = sitemap.readers(document_name.as_str(), &self.package.groups);
+            if document_readers.is_empty() {
+                return Ok(true);
+            }
+            return fpm::user_group::belongs_to(
+                self,
+                document_readers.as_slice(),
+                access_identities.iter().collect_vec().as_slice(),
+            );
+        }
+        Ok(true)
+    }
+
+    pub(crate) async fn can_write(
+        &self,
+        req: &actix_web::HttpRequest,
+        document_path: &str,
+    ) -> fpm::Result<bool> {
+        use itertools::Itertools;
+        let document_name = self.document_name_with_default(document_path);
+        let access_identities =
+            fpm::user_group::access_identities(self, req, &document_name, false).await?;
+
+        if let Some(sitemap) = &self.package.sitemap {
+            // TODO: This can be buggy in case of: if groups are used directly in sitemap are foreign groups
+            let document_writers = sitemap.writers(document_name.as_str(), &self.package.groups);
+            return fpm::user_group::belongs_to(
+                self,
+                document_writers.as_slice(),
+                access_identities.iter().collect_vec().as_slice(),
+            );
+        }
+
+        Ok(false)
     }
 }
 
@@ -1560,398 +1785,9 @@ impl Package {
         resp
     }
 
-    pub async fn get_assets_doc(
-        &self,
-        config: &fpm::Config,
-        base_url: &str,
-    ) -> fpm::Result<String> {
+    pub async fn get_assets_doc(&self) -> fpm::Result<String> {
         // Virtual document that contains the asset information about the package
-        use itertools::Itertools;
-        let all_docs = config.get_files(self).await?;
-        let all_files = all_docs.into_iter().filter_map(|file_instance| {
-            let id = file_instance.get_id();
-            if id.eq("FPM.ftd") {
-                None
-            } else {
-                Some((Path::new(id.as_str()), file_instance))
-            }
-        });
-        let mut top = Dir::new("root", "root", "/", None);
-        for (path, file_ins) in all_files {
-            build_tree(&mut top, &path.parts, 0, Some(file_ins));
-        }
-        let mut all_extensions: Vec<String> = vec![];
-        let (generated_records, generated_values) = build_record_values(
-            &top,
-            &mut all_extensions,
-            self.name.as_str(),
-            config.package.name.as_str(),
-            base_url,
-        );
-        let (font_record, fonts) = self
-            .fonts
-            .iter()
-            .unique_by(|font| font.name.as_str())
-            .collect_vec()
-            .iter()
-            .fold(
-                (
-                    String::from("-- record font:"),
-                    String::from("-- font fonts:"),
-                ),
-                |(record_accumulator, instance_accumulator), font| {
-                    (
-                        format!(
-                            "{pre}\nstring {font_var_name}:",
-                            pre = record_accumulator,
-                            font_var_name = font.name.as_str(),
-                        ),
-                        format!(
-                            "{pre}\n{font_var_name}: {font_var_val}",
-                            pre = instance_accumulator,
-                            font_var_name = font.name.as_str(),
-                            font_var_val = font.html_name(self.name.as_str())
-                        ),
-                    )
-                },
-            );
-        return Ok(format!(
-            indoc::indoc! {"
-                {generated_records}\n
-                {generated_values}
-                {font_record}
-                {fonts}
-            "},
-            generated_records = generated_records,
-            generated_values = generated_values,
-            font_record = font_record,
-            fonts = fonts
-        ));
-
-        #[derive(Debug)]
-        struct Path {
-            parts: Vec<String>,
-        }
-        impl Path {
-            pub fn new(path: &str) -> Path {
-                Path {
-                    parts: path
-                        .to_string()
-                        .split(std::path::MAIN_SEPARATOR)
-                        .map(|s| s.to_string())
-                        .collect(),
-                }
-            }
-        }
-
-        #[derive(Debug, Clone)]
-        struct Dir {
-            name: String,
-            full_path: String,
-            dir_path: String,
-            file_instance: Option<fpm::File>,
-            children: Vec<Dir>,
-        }
-
-        impl Dir {
-            fn new(
-                name: &str,
-                full_path: &str,
-                dir_path: &str,
-                file_instance: Option<fpm::File>,
-            ) -> Dir {
-                Dir {
-                    name: name.to_string(),
-                    full_path: full_path.to_string(),
-                    dir_path: dir_path.to_string(),
-                    file_instance,
-                    children: Vec::<Dir>::new(),
-                }
-            }
-
-            fn find_child(&mut self, name: &str) -> Option<&mut Dir> {
-                for c in self.children.iter_mut() {
-                    if c.name == name {
-                        return Some(c);
-                    }
-                }
-                None
-            }
-
-            fn add_child<T>(&mut self, leaf: T) -> &mut Self
-            where
-                T: Into<Dir>,
-            {
-                self.children.push(leaf.into());
-                self
-            }
-
-            fn is_leaf(&self) -> bool {
-                self.children.is_empty()
-            }
-
-            fn full_path_to_key(&self) -> String {
-                get_sanitized_string(self.full_path.as_str())
-            }
-            fn name_and_extension(&self) -> (&str, Option<&str>) {
-                if let Some((name, ext)) = self.name.as_str().rsplit_once('.') {
-                    let name = if name.is_empty() { "DOT" } else { name };
-                    (name, Some(ext))
-                } else {
-                    // File without extension
-                    (self.name.as_str(), None)
-                }
-            }
-        }
-
-        fn get_sanitized_string(input: &str) -> String {
-            let path = input.trim_start_matches('.').trim();
-            path.chars()
-                .map(|x| match x {
-                    '/' => '-',
-                    '\\' => '-',
-                    '.' => '-',
-                    '_' => '-',
-                    _ => x,
-                })
-                .collect()
-        }
-
-        fn build_tree(node: &mut Dir, parts: &[String], depth: usize, file_ins: Option<fpm::File>) {
-            if depth < parts.len() {
-                let item = &parts[depth];
-                let full_path = &parts[..depth + 1].join("/");
-                let dir_path = &parts[..depth].join("/");
-
-                let dir = match node.find_child(item) {
-                    Some(d) => d,
-                    None => {
-                        let d = Dir::new(
-                            item,
-                            full_path.as_str(),
-                            dir_path.as_str(),
-                            if let Some(f) = file_ins.clone() {
-                                if full_path.as_str().eq(&f
-                                    .get_id()
-                                    .as_str()
-                                    .replace(std::path::MAIN_SEPARATOR, "/"))
-                                {
-                                    Some(f)
-                                } else {
-                                    None
-                                }
-                            } else {
-                                None
-                            },
-                        );
-                        node.add_child(d);
-                        match node.find_child(item) {
-                            Some(d2) => d2,
-                            None => panic!("Got here!"),
-                        }
-                    }
-                };
-                build_tree(dir, parts, depth + 1, file_ins);
-            }
-        }
-
-        fn build_record_values(
-            node: &Dir,
-            found_extensions: &mut Vec<String>,
-            package_name: &str,
-            current_package_name: &str,
-            base_url: &str,
-        ) -> (String, String) {
-            // Strip base url's end slashes. The code takes care of the URL evaluation
-            let base_url = base_url.trim_end_matches('/');
-            let mut resp_records = String::new();
-            let mut resp_values = String::new();
-            let mut root_record = String::from("-- record all-files:");
-            let mut root_record_instance = String::from("-- all-files files:");
-            let mut root_record_page = String::from("-- record all-pages:");
-            let mut root_record_instance_page = String::from("-- all-pages pages:");
-            // First combine all the children by name
-            let mut named_children: std::collections::HashMap<String, Vec<Dir>> =
-                std::collections::HashMap::new();
-
-            for child_node in node.children.clone().into_iter() {
-                // let child_node = child_node.deref();
-                let (child_name, _) = child_node.name_and_extension();
-                let child_key = get_sanitized_string(
-                    format!(
-                        "{dir_path}{slash}{child_name}",
-                        dir_path = child_node.dir_path.as_str(),
-                        slash = if child_node.dir_path.is_empty() {
-                            ""
-                        } else {
-                            "/"
-                        }
-                    )
-                    .as_str(),
-                );
-                if let Some(ins) = named_children.get_mut(&child_key) {
-                    ins.push(child_node);
-                } else {
-                    named_children.insert(child_key, vec![child_node]);
-                };
-            }
-            for (key_name, children) in named_children.iter() {
-                let mut named_child_record = format!("-- record file-record-{key_name}:");
-                let mut named_child_instance =
-                    format!("-- file-record-{key_name} file-record-instance-{key_name}:");
-                let mut named_child_record_page = format!("-- record page-record-{key_name}:");
-                let mut named_child_instance_page =
-                    format!("-- page-record-{key_name} page-record-instance-{key_name}:");
-                if node.name.as_str().eq("root") {
-                    root_record = format!("{root_record}\nfile-record-{key_name} {key_name}:");
-                    root_record_instance = format!(
-                        "{root_record_instance}\n{key_name}: $file-record-instance-{key_name}"
-                    );
-                    root_record_page =
-                        format!("{root_record_page}\npage-record-{key_name} {key_name}:");
-                    root_record_instance_page = format!(
-                        "{root_record_instance_page}\n{key_name}: $page-record-instance-{key_name}"
-                    );
-                }
-                for child in children {
-                    let (child_name, child_ext) = child.name_and_extension();
-
-                    if child.is_leaf() {
-                        let attribute_name: &str = if let Some(ext) = child_ext {
-                            // found_extensions.push(ext.to_string());
-                            ext
-                        } else {
-                            "without-extension"
-                        };
-                        let page_record_part = format!("string {attribute_name}-page:");
-                        // TODO: The URL for the generated page to be fixed
-                        let page_instance_part = format!(
-                            "{attribute_name}-page: {child_dir_path}/{child_name}",
-                            child_dir_path = child.dir_path
-                        );
-                        let (append_page, is_static_copied, attribute_type) =
-                            match child.file_instance.as_ref().unwrap() {
-                                fpm::File::Image(_) => {
-                                    // In case markdown, append the md-page attribute directly
-                                    (true, true, "ftd.image-src")
-                                }
-                                fpm::File::Markdown(_) | fpm::File::Code(_) => {
-                                    (true, true, "string")
-                                }
-                                fpm::File::Ftd(_) => (false, false, "string"),
-                                _ => (false, true, "string"),
-                            };
-                        if append_page {
-                            named_child_record =
-                                format!("{named_child_record}\n{page_record_part}");
-                            named_child_instance =
-                                format!("{named_child_instance}\n{page_instance_part}");
-                        }
-                        if attribute_type.eq("ftd.image-src") {
-                            let dark_mode_file_name = format!("{child_name}-dark.{attribute_name}");
-                            let dark_mode_asset = node
-                                .children
-                                .iter()
-                                .find(|c| c.name.eq(dark_mode_file_name.as_str()));
-                            let image_src = format!(
-                                indoc::indoc! {"
-                                    -- ftd.image-src file-leaf-instance-{child_record_instance}:
-                                    light: {base_url}/-/{package_name}/{child_full_path}
-                                    dark: {base_url}/-/{package_name}/{dark_mode_file_path}"
-                                },
-                                base_url = base_url,
-                                package_name = package_name,
-                                child_record_instance = child.full_path_to_key(),
-                                child_full_path = child.full_path.as_str(),
-                                dark_mode_file_path = if let Some(dark_asset) = dark_mode_asset {
-                                    dark_asset.full_path.as_str()
-                                } else {
-                                    child.full_path.as_str()
-                                },
-                            );
-                            resp_values = format!("{image_src}\n{resp_values}",);
-                        } else {
-                            resp_values = format!(
-                                "-- {attribute_type} file-leaf-instance-{child_record_instance}: {base_url}/{static_dir_prefix}{child_instance_path}\n{resp_values}",
-                                static_dir_prefix = if is_static_copied { format!("-/{package_name}/")} else if !package_name.eq(current_package_name){format!("{package_name}/")} else {String::new()},
-                                child_record_instance = child.full_path_to_key(),
-                                child_instance_path = child.full_path
-                            );
-                        }
-
-                        resp_values = format!(
-                            "-- string page-leaf-instance-{child_record_instance}: {base_url}/{static_dir_prefix}{child_instance_path}\n{resp_values}",
-                            static_dir_prefix = if !package_name.eq(current_package_name){format!("{package_name}/")} else {String::new()},
-                            child_record_instance = child.full_path_to_key(),
-                            child_instance_path = child.full_path
-                        );
-
-                        named_child_record =
-                            format!("{named_child_record}\n{attribute_type} {attribute_name}:");
-                        named_child_instance = format!(
-                            "{named_child_instance}\n{attribute_name}: $file-leaf-instance-{child_record_instance}",
-                            child_record_instance = child.full_path_to_key()
-                        );
-
-                        named_child_record_page =
-                            format!("{named_child_record_page}\nstring {attribute_name}:");
-                        named_child_instance_page = format!(
-                            "{named_child_instance_page}\n{attribute_name}: $page-leaf-instance-{child_record_instance}",
-                            child_record_instance = child.full_path_to_key()
-                        );
-                    } else {
-                        let (child_records, child_values) = build_record_values(
-                            child,
-                            found_extensions,
-                            package_name,
-                            current_package_name,
-                            base_url,
-                        );
-                        resp_records = format!("{child_records}\n{resp_records}");
-                        resp_values = format!("{child_values}\n{resp_values}");
-
-                        for sub_child in child.children.iter() {
-                            let (sub_child_name, _) = sub_child.name_and_extension();
-
-                            let sub_child_key = get_sanitized_string(
-                                format!(
-                                    "{dir_path}{slash}{sub_child_name}",
-                                    dir_path = sub_child.dir_path.as_str(),
-                                    slash = if sub_child.dir_path.is_empty() {
-                                        ""
-                                    } else {
-                                        "/"
-                                    }
-                                )
-                                .as_str(),
-                            );
-                            named_child_record = format!(
-                                "{named_child_record}\nfile-record-{sub_child_key} {sub_child_name}:",
-                            );
-                            named_child_instance = format!(
-                                "{named_child_instance}\n{sub_child_name}: $file-record-instance-{sub_child_key}",
-                            );
-                            named_child_record_page = format!(
-                                "{named_child_record_page}\npage-record-{sub_child_key} {sub_child_name}:",
-                            );
-                            named_child_instance_page = format!(
-                                "{named_child_instance_page}\n{sub_child_name}: $page-record-instance-{sub_child_key}",
-                            );
-                        }
-                    }
-                }
-                resp_records =
-                    format!("{resp_records}\n{named_child_record}\n{named_child_record_page}");
-                resp_values =
-                    format!("{resp_values}\n{named_child_instance}\n{named_child_instance_page}");
-            }
-            if node.name.as_str().eq("root") {
-                resp_records = format!("{resp_records}\n{root_record}\n{root_record_page}");
-                resp_values =
-                    format!("{resp_values}\n{root_record_instance}\n{root_record_instance_page}");
-            }
-            (resp_records, resp_values)
-        }
+        Ok(self.get_font_ftd().unwrap_or_default())
     }
 
     pub(crate) async fn get_fpm(&self) -> fpm::Result<String> {
