@@ -1,8 +1,16 @@
-#[tokio::main]
-async fn main() -> fpm::Result<()> {
+fn main() -> fpm::Result<()> {
+    tokio::runtime::Builder::new_multi_thread()
+        .enable_all()
+        .build()
+        .unwrap()
+        .block_on(async_main())
+}
+
+async fn async_main() -> fpm::Result<()> {
+    use colored::Colorize;
+
     let matches = app(authors(), version()).get_matches();
 
-    // Block of code to run when start-project subcommand is used
     if let Some(project) = matches.subcommand_matches("start-project") {
         // project-name => required field (any package Url or standard project name)
         let name = project.value_of("package-name").unwrap();
@@ -12,24 +20,17 @@ async fn main() -> fpm::Result<()> {
         return Ok(());
     }
 
-    // Serve block moved up
     if let Some(mark) = matches.subcommand_matches("serve") {
-        let port = mark
-            .value_of("port")
-            .map_or(mark.value_of("positional_port"), Some)
-            .map(|p| {
-                p.parse::<u16>()
-                    .unwrap_or_else(|_| panic!("provided port {} is wrong", p))
-            });
-
-        let identities = mark.value_of("identities").map(|x| x.to_string());
+        let port = mark.value_of("port").map(|p| match p.parse::<u16>() {
+            Ok(v) => v,
+            Err(_) => {
+                eprintln!("Provided port {} is not a valid port.", p.to_string().red());
+                std::process::exit(1);
+            }
+        });
 
         let bind = mark.value_of("bind").unwrap_or("127.0.0.1").to_string();
-        tokio::task::spawn_blocking(move || {
-            fpm::fpm_serve(bind.as_str(), port, identities).expect("http service error");
-        })
-        .await
-        .expect("Thread spawn error");
+        fpm::fpm_serve(bind.as_str(), port).await?;
         return Ok(());
     }
 
@@ -44,13 +45,34 @@ async fn main() -> fpm::Result<()> {
         fpm::update(&config).await?;
     }
 
+    if let Some(edit) = matches.subcommand_matches("edit") {
+        fpm::edit(
+            &config,
+            edit.value_of("file").unwrap(),
+            edit.value_of("cr").unwrap(),
+        )
+        .await?;
+        return Ok(());
+    }
+
     if let Some(add) = matches.subcommand_matches("add") {
         fpm::add(&config, add.value_of("file").unwrap(), add.value_of("cr")).await?;
         return Ok(());
     }
 
     if let Some(rm) = matches.subcommand_matches("rm") {
-        fpm::rm(&config, rm.value_of("file").unwrap()).await?;
+        fpm::rm(&config, rm.value_of("file").unwrap(), rm.value_of("cr")).await?;
+        return Ok(());
+    }
+
+    if let Some(merge) = matches.subcommand_matches("merge") {
+        fpm::merge(
+            &config,
+            merge.value_of("src"),
+            merge.value_of("dest").unwrap(),
+            merge.value_of("file"),
+        )
+        .await?;
         return Ok(());
     }
 
@@ -92,8 +114,9 @@ async fn main() -> fpm::Result<()> {
         let source = status.value_of("source");
         fpm::sync_status(&config, source).await?;
     }
-    if matches.subcommand_matches("create-cr").is_some() {
-        fpm::create_cr(&config).await?;
+    if let Some(create_cr) = matches.subcommand_matches("create-cr") {
+        let title = create_cr.value_of("title");
+        fpm::create_cr(&config, title).await?;
     }
     if let Some(close_cr) = matches.subcommand_matches("close-cr") {
         let cr = close_cr.value_of("cr").unwrap();
@@ -145,15 +168,14 @@ async fn main() -> fpm::Result<()> {
     Ok(())
 }
 
-fn app(authors: &'static str, version: &'static str) -> clap::App<'static, 'static> {
+fn app(authors: &'static str, version: &'static str) -> clap::App<'static> {
     clap::App::new("fpm: FTD Package Manager")
         .version(version)
         .author(authors)
         .setting(clap::AppSettings::ArgRequiredElseHelp)
         .arg(
             clap::Arg::with_name("verbose")
-                .short("v")
-                .multiple(true)
+                .short('v')
                 .help("Sets the level of verbosity"),
         )
         .arg(
@@ -176,7 +198,7 @@ fn app(authors: &'static str, version: &'static str) -> clap::App<'static, 'stat
                 )
                 .arg(
                     clap::Arg::with_name("package-path")
-                        .short("p")
+                        .short('p')
                         .long("path")
                         .takes_value(true)
                         .help("Package path (relative)")
@@ -203,7 +225,7 @@ fn app(authors: &'static str, version: &'static str) -> clap::App<'static, 'stat
                 .arg(
                     clap::Arg::with_name("verbose")
                         .long("verbose")
-                        .short("v")
+                        .short('v')
                         .takes_value(false)
                         .required(false),
                 )
@@ -228,6 +250,15 @@ fn app(authors: &'static str, version: &'static str) -> clap::App<'static, 'stat
                 .version(env!("CARGO_PKG_VERSION")),
         )
         .subcommand(
+            clap::SubCommand::with_name("edit")
+                .about("Edit a file in CR workspace")
+                .args(&[
+                    clap::Arg::with_name("file").required(true),
+                    clap::Arg::with_name("cr").long("--cr").takes_value(true).required(true),
+                ])
+                .version(env!("CARGO_PKG_VERSION")),
+        )
+        .subcommand(
             clap::SubCommand::with_name("add")
                 .about("Adds a file in workspace")
                 .args(&[
@@ -239,7 +270,20 @@ fn app(authors: &'static str, version: &'static str) -> clap::App<'static, 'stat
         .subcommand(
             clap::SubCommand::with_name("rm")
                 .about("Removes a file in workspace")
-                .arg(clap::Arg::with_name("file").required(true))
+                .args(&[
+                    clap::Arg::with_name("file").required(true),
+                    clap::Arg::with_name("cr").long("--cr").takes_value(true),
+                ])
+                .version(env!("CARGO_PKG_VERSION")),
+        )
+        .subcommand(
+            clap::SubCommand::with_name("merge")
+                .about("Merge two manifests together")
+                .args(&[
+                    clap::Arg::with_name("src").long("--src").takes_value(true),
+                    clap::Arg::with_name("dest").long("--dest").takes_value(true).required(true),
+                    clap::Arg::with_name("file"),
+                ])
                 .version(env!("CARGO_PKG_VERSION")),
         )
         .subcommand(
@@ -273,6 +317,7 @@ fn app(authors: &'static str, version: &'static str) -> clap::App<'static, 'stat
         )
         .subcommand(
             clap::SubCommand::with_name("create-cr")
+                .arg(clap::Arg::with_name("title"))
                 .about("Create a Change Request")
                 .version(env!("CARGO_PKG_VERSION")),
         )
@@ -291,7 +336,7 @@ fn app(authors: &'static str, version: &'static str) -> clap::App<'static, 'stat
             clap::SubCommand::with_name("diff")
                 .args(&[
                     clap::Arg::with_name("source").multiple(true),
-                    clap::Arg::with_name("all").long("--all").short("a"),
+                    clap::Arg::with_name("all").long("--all").short('a'),
                 ])
                 .about("Show un-synced changes to files in this fpm package")
                 .version(env!("CARGO_PKG_VERSION")),
@@ -348,38 +393,51 @@ fn app(authors: &'static str, version: &'static str) -> clap::App<'static, 'stat
                 .about("Remove a tracking relation between two files")
                 .version(env!("CARGO_PKG_VERSION")),
         )
-        .subcommand(
-            clap::SubCommand::with_name("serve")
-                .arg(clap::Arg::with_name("port")
-                        .required_unless("positional_port")
-                        .required(false)
-                        .help("Specify the port to serve on")
-                )
-                .arg(clap::Arg::with_name("positional_port")
-                        .long("--port")
-                        .required_unless("bind")
-                        .takes_value(true)
-                        .required(false)
-                        .help("Specify the port to serve on")
-                )
-                .arg(clap::Arg::with_name("bind")
-                        .long("--bind")
-                        .takes_value(true)
-                        .required(false)
-                        .help("Specify the bind address to serve on")
-                )
-                .about("Create an http server and serves static files")
-                .version(env!("CARGO_PKG_VERSION")),
+        .subcommand(sub_command_serve())
+}
+
+fn sub_command_serve() -> clap::App<'static> {
+    let serve = clap::SubCommand::with_name("serve")
+        .arg(
+            clap::Arg::with_name("port")
+                .takes_value(true)
+                .help("Specify the port to serve on"),
         )
+        .arg(
+            clap::Arg::with_name("bind")
+                .long("--bind")
+                .takes_value(true)
+                .help("Specify the bind address to serve on"),
+        );
+
+    if cfg!(feature = "remote") {
+        serve
+    } else {
+        serve
+            .arg(
+                clap::Arg::with_name("identities")
+                    .long("--identities")
+                    .takes_value(true)
+                    .required(false)
+                    .help(
+                        "Http request identities, fpm allows these identities to access documents",
+                    ),
+            )
+            .about("Create an http server and serves static files")
+            .version(env!("CARGO_PKG_VERSION"))
+    }
 }
 
 pub fn version() -> &'static str {
     if std::env::args().any(|e| e == "--test") {
         env!("CARGO_PKG_VERSION")
     } else {
-        Box::leak(
-            format!("{} [{}]", env!("CARGO_PKG_VERSION"), env!("VERGEN_GIT_SHA")).into_boxed_str(),
-        )
+        match option_env!("GITHUB_SHA") {
+            Some(sha) => {
+                Box::leak(format!("{} [{}]", env!("CARGO_PKG_VERSION"), sha).into_boxed_str())
+            }
+            None => env!("CARGO_PKG_VERSION"),
+        }
     }
 }
 
