@@ -364,7 +364,7 @@ pub enum ParseError {
     },
 }
 
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 enum ParsingState {
     WaitingForSection,
     ParsingSection,
@@ -917,21 +917,6 @@ impl Sitemap {
         }
     }
 
-    fn ids_matches(id1: &str, id2: &str) -> bool {
-        return strip_id(id1).eq(&strip_id(id2));
-
-        fn strip_id(id: &str) -> String {
-            let id = id
-                .trim()
-                .replace("/index.html", "/")
-                .replace("index.html", "/");
-            if id.eq("/") {
-                return id;
-            }
-            id.trim_matches('/').to_string()
-        }
-    }
-
     pub(crate) fn get_sitemap_by_id(&self, id: &str) -> Option<SiteMapCompat> {
         let mut sections = vec![];
         let mut subsections = vec![];
@@ -943,7 +928,7 @@ impl Sitemap {
         for (idx, section) in self.sections.iter().enumerate() {
             index = idx;
 
-            if fpm::sitemap::Sitemap::ids_matches(section.id.as_str(), id) {
+            if fpm::utils::ids_matches(section.id.as_str(), id) {
                 subsections = section
                     .subsections
                     .iter()
@@ -951,14 +936,14 @@ impl Sitemap {
                     .filter(|v| {
                         let active =
                             v.id.as_ref()
-                                .map(|v| fpm::sitemap::Sitemap::ids_matches(v, id))
+                                .map(|v| fpm::utils::ids_matches(v, id))
                                 .unwrap_or(false);
                         active || !v.skip
                     })
                     .map(|v| {
                         let active =
                             v.id.as_ref()
-                                .map(|v| fpm::sitemap::Sitemap::ids_matches(v, id))
+                                .map(|v| fpm::utils::ids_matches(v, id))
                                 .unwrap_or(false);
                         let toc = TocItemCompat::new(
                             v.id.clone(),
@@ -985,7 +970,7 @@ impl Sitemap {
                     .filter(|s| !s.skip)
                     .find_or_first(|v| {
                         v.id.as_ref()
-                            .map(|v| fpm::sitemap::Sitemap::ids_matches(v, id))
+                            .map(|v| fpm::utils::ids_matches(v, id))
                             .unwrap_or(false)
                     })
                     .or_else(|| section.subsections.first())
@@ -1093,7 +1078,7 @@ impl Sitemap {
                     && subsection
                         .id
                         .as_ref()
-                        .map(|v| fpm::sitemap::Sitemap::ids_matches(v, id))
+                        .map(|v| fpm::utils::ids_matches(v, id))
                         .unwrap_or(false)
                 {
                     let (toc_list, current_toc) = get_all_toc(subsection.toc.as_slice(), id);
@@ -1195,7 +1180,7 @@ impl Sitemap {
             for toc_item in toc.iter() {
                 let (is_open, children) =
                     get_toc_by_id_(id, toc_item.children.as_slice(), current_page);
-                let is_active = fpm::sitemap::Sitemap::ids_matches(toc_item.id.as_str(), id);
+                let is_active = fpm::utils::ids_matches(toc_item.id.as_str(), id);
                 let current_toc = {
                     let mut current_toc = TocItemCompat::new(
                         Some(get_url(toc_item.id.as_str()).to_string()),
@@ -1213,7 +1198,7 @@ impl Sitemap {
                 };
 
                 if current_page.is_none() {
-                    found_here = fpm::sitemap::Sitemap::ids_matches(toc_item.id.as_str(), id);
+                    found_here = fpm::utils::ids_matches(toc_item.id.as_str(), id);
                     if found_here {
                         let mut current_toc = current_toc.clone();
                         if let Some(ref title) = toc_item.nav_title {
@@ -1247,7 +1232,7 @@ impl Sitemap {
         id: &str,
     ) -> Option<std::collections::BTreeMap<String, String>> {
         for section in self.sections.iter() {
-            if fpm::sitemap::Sitemap::ids_matches(section.id.as_str(), id) {
+            if fpm::utils::ids_matches(section.id.as_str(), id) {
                 return Some(section.extra_data.to_owned());
             }
             if let Some(data) = get_extra_data_from_subsections(id, section.subsections.as_slice())
@@ -1265,7 +1250,7 @@ impl Sitemap {
         ) -> Option<std::collections::BTreeMap<String, String>> {
             for subsection in subsections {
                 if subsection.visible
-                    && fpm::sitemap::Sitemap::ids_matches(
+                    && fpm::utils::ids_matches(
                         subsection.id.as_ref().unwrap_or(&"".to_string()),
                         id,
                     )
@@ -1286,7 +1271,7 @@ impl Sitemap {
             toc: &[TocItem],
         ) -> Option<std::collections::BTreeMap<String, String>> {
             for toc_item in toc {
-                if fpm::sitemap::Sitemap::ids_matches(toc_item.id.as_str(), id) {
+                if fpm::utils::ids_matches(toc_item.id.as_str(), id) {
                     return Some(toc_item.extra_data.to_owned());
                 }
                 if let Some(data) = get_extra_data_from_toc(id, toc_item.children.as_slice()) {
@@ -1300,7 +1285,9 @@ impl Sitemap {
     }
 
     /// This function will return all the readers and readers which are inherited from parent
-    // TODO: All writers are also readers
+
+    // TODO: need to handle special reader: everyone, writer: everyone
+    // readers function should return Vec<UserGroup> or Everyone
     pub fn readers<'a>(
         &self,
         doc_path: &str,
@@ -1316,6 +1303,7 @@ impl Sitemap {
                 .iter()
                 .unique()
                 .filter_map(|g| groups.get(g))
+                .chain(self.writers(doc_path, groups))
                 .collect();
         }
 
