@@ -1,7 +1,7 @@
 use itertools::Itertools;
 use sha2::Digest;
 
-#[derive(Clone, PartialEq, Debug)]
+#[derive(Clone, PartialEq, Eq, Debug)]
 pub enum Status {
     Conflict(i32),
     NoConflict,
@@ -28,7 +28,7 @@ impl Status {
     }
 }
 
-#[derive(Clone)]
+#[derive(Debug, Clone)]
 pub enum FileStatus {
     Add {
         path: String,
@@ -49,11 +49,12 @@ pub enum FileStatus {
     Uptodate {
         path: String,
         version: i32,
+        content: Vec<u8>,
     },
 }
 
 impl FileStatus {
-    fn is_conflicted(&self) -> bool {
+    pub(crate) fn is_conflicted(&self) -> bool {
         let status = match self {
             FileStatus::Add { status, .. }
             | FileStatus::Update { status, .. }
@@ -96,14 +97,19 @@ impl FileStatus {
         }
     }
 
-    pub(crate) fn sync_request(self) -> Option<fpm::apis::sync2::SyncRequestFile> {
+    pub(crate) fn sync_request(
+        self,
+        src_cr: Option<usize>,
+    ) -> Option<fpm::apis::sync2::SyncRequestFile> {
         if self.is_conflicted() {
             return None;
         }
         Some(match self {
-            FileStatus::Add { path, content, .. } => {
-                fpm::apis::sync2::SyncRequestFile::Add { path, content }
-            }
+            FileStatus::Add { path, content, .. } => fpm::apis::sync2::SyncRequestFile::Add {
+                path,
+                content,
+                src_cr,
+            },
             FileStatus::Update {
                 path,
                 content,
@@ -113,10 +119,13 @@ impl FileStatus {
                 path,
                 content,
                 version,
+                src_cr,
             },
-            FileStatus::Delete { path, version, .. } => {
-                fpm::apis::sync2::SyncRequestFile::Delete { path, version }
-            }
+            FileStatus::Delete { path, version, .. } => fpm::apis::sync2::SyncRequestFile::Delete {
+                path,
+                version,
+                src_cr,
+            },
             FileStatus::Uptodate { .. } => return None,
         })
     }
@@ -124,12 +133,8 @@ impl FileStatus {
 
 impl fpm::Config {
     pub(crate) async fn get_files_status(&self) -> fpm::Result<Vec<FileStatus>> {
-        let file_list = self.read_workspace().await?;
         let mut workspace: std::collections::BTreeMap<String, fpm::workspace::WorkspaceEntry> =
-            file_list
-                .iter()
-                .map(|v| (v.filename.to_string(), v.clone()))
-                .collect();
+            self.get_workspace_map().await?;
         let changed_files = self.get_files_status_with_workspace(&mut workspace).await?;
         self.write_workspace(workspace.into_values().collect_vec().as_slice())
             .await?;
@@ -140,19 +145,41 @@ impl fpm::Config {
         &self,
         workspace: &mut std::collections::BTreeMap<String, fpm::workspace::WorkspaceEntry>,
     ) -> fpm::Result<Vec<FileStatus>> {
-        let mut changed_files = self.get_files_status_wrt_workspace(workspace).await?;
-        self.get_files_status_wrt_remote_latest(&mut changed_files, workspace)
+        let mut changed_files: std::collections::BTreeMap<String, FileStatus> = self
+            .get_files_status_wrt_workspace(workspace)
+            .await?
+            .into_iter()
+            .map(|v| (v.get_file_path(), v))
+            .collect();
+        self.get_files_status_wrt_remote_manifest(&mut changed_files, workspace)
             .await?;
-        Ok(changed_files)
+        Ok(changed_files.into_values().collect_vec())
     }
+
+    /*async fn update_workspace_using_cr_track(
+        &self,
+        cr_workspace: &mut fpm::workspace::CRWorkspace,
+    ) -> fpm::Result<Vec<FileStatus>> {
+        let cr_tracking_infos = self.get_cr_tracking_info(cr_workspace.cr).await?;
+        for tracking_info in cr_tracking_infos {
+            if cr_workspace.workspace.contains_key(&tracking_info.filename) {
+                continue;
+            }
+            cr_workspace.workspace.insert(
+                tracking_info.filename,
+                tracking_info.into_workspace_entry(cr_workspace.cr),
+            )
+        }
+
+        Ok(())
+    }*/
 
     async fn get_files_status_wrt_workspace(
         &self,
         workspace: &std::collections::BTreeMap<String, fpm::workspace::WorkspaceEntry>,
     ) -> fpm::Result<Vec<FileStatus>> {
-        let files = workspace.values().collect_vec();
         let mut changed_files = vec![];
-        for workspace_entry in files {
+        for (filename, workspace_entry) in workspace {
             let version = if let Some(version) = workspace_entry.version {
                 version
             } else {
@@ -181,12 +208,13 @@ impl fpm::Config {
 
             let content =
                 tokio::fs::read(self.root.join(workspace_entry.filename.as_str())).await?;
-            let history_path = self.history_path(workspace_entry.filename.as_str(), version);
+            let history_path = self.history_path(filename.as_str(), version);
             let history_content = tokio::fs::read(history_path).await?;
             if sha2::Sha256::digest(&content).eq(&sha2::Sha256::digest(&history_content)) {
                 changed_files.push(FileStatus::Uptodate {
                     path: workspace_entry.filename.to_string(),
                     version,
+                    content,
                 });
                 continue;
             }
@@ -200,14 +228,33 @@ impl fpm::Config {
         Ok(changed_files)
     }
 
-    async fn get_files_status_wrt_remote_latest(
+    pub(crate) async fn get_files_status_wrt_remote_manifest(
         &self,
-        files: &mut Vec<FileStatus>,
+        files: &mut std::collections::BTreeMap<String, FileStatus>,
         workspace: &mut std::collections::BTreeMap<String, fpm::workspace::WorkspaceEntry>,
     ) -> fpm::Result<()> {
-        let mut remove_files = vec![];
-        let server_latest = self.get_latest_file_edits().await?;
-        for (index, file) in files.iter_mut().enumerate() {
+        let remote_manifest = self.get_remote_manifest(true).await?;
+        let (already_added_files, already_removed_files) = self
+            .get_files_status_wrt_manifest(files, &remote_manifest)
+            .await?;
+        for file_name in already_removed_files {
+            workspace.remove(file_name.as_str());
+        }
+        for workspace_entry in already_added_files {
+            workspace.insert(workspace_entry.filename.to_string(), workspace_entry);
+        }
+        Ok(())
+    }
+
+    async fn get_files_status_wrt_manifest(
+        &self,
+        files: &mut std::collections::BTreeMap<String, FileStatus>,
+        manifest: &std::collections::BTreeMap<String, fpm::history::FileEdit>,
+    ) -> fpm::Result<(Vec<fpm::workspace::WorkspaceEntry>, Vec<String>)> {
+        let mut already_removed_files = vec![];
+        let mut already_added_files = vec![];
+        let mut uptodate_files = vec![];
+        for (filename, file) in files.iter_mut() {
             match file {
                 FileStatus::Uptodate { .. } => {
                     continue;
@@ -217,7 +264,7 @@ impl fpm::Config {
                     content,
                     status,
                 } => {
-                    let server_version = if let Some(file_edit) = server_latest.get(path) {
+                    let server_version = if let Some(file_edit) = manifest.get(path) {
                         if file_edit.is_deleted() {
                             continue;
                         }
@@ -228,16 +275,13 @@ impl fpm::Config {
                     let history_path = self.history_path(path, server_version);
                     let history_content = tokio::fs::read(history_path).await?;
                     if sha2::Sha256::digest(content).eq(&sha2::Sha256::digest(history_content)) {
-                        workspace.insert(
-                            path.to_string(),
-                            fpm::workspace::WorkspaceEntry {
-                                filename: path.to_string(),
-                                deleted: None,
-                                version: Some(server_version),
-                                cr: None,
-                            },
-                        );
-                        remove_files.push(index);
+                        already_added_files.push(fpm::workspace::WorkspaceEntry {
+                            filename: path.to_string(),
+                            deleted: None,
+                            version: Some(server_version),
+                            cr: None,
+                        });
+                        uptodate_files.push(filename.clone())
                     } else {
                         *status = Status::CloneAddedRemoteAdded(server_version);
                     }
@@ -248,7 +292,7 @@ impl fpm::Config {
                     version,
                     status,
                 } => {
-                    let server_file_edit = if let Some(file_edit) = server_latest.get(path) {
+                    let server_file_edit = if let Some(file_edit) = manifest.get(path) {
                         file_edit
                     } else {
                         continue;
@@ -286,7 +330,7 @@ impl fpm::Config {
                         .merge(&ancestor_content, &ours_content, &theirs_content)
                     {
                         Ok(data) => {
-                            tokio::fs::write(path, &data).await?;
+                            fpm::utils::update(self.root.join(filename), data.as_bytes()).await?;
                             *content = data.as_bytes().to_vec();
                             *version = server_file_edit.version;
                         }
@@ -301,16 +345,16 @@ impl fpm::Config {
                     version,
                     status,
                 } => {
-                    let server_file_edit = if let Some(server_file_edit) = server_latest.get(path) {
+                    let server_file_edit = if let Some(server_file_edit) = manifest.get(path) {
                         server_file_edit
                     } else {
-                        remove_files.push(index);
-                        workspace.remove(path);
+                        already_removed_files.push(filename.clone());
+                        uptodate_files.push(filename.clone());
                         continue;
                     };
                     if server_file_edit.is_deleted() {
-                        remove_files.push(index);
-                        workspace.remove(path);
+                        already_removed_files.push(filename.clone());
+                        uptodate_files.push(filename.clone());
                         continue;
                     }
                     if !server_file_edit.version.eq(version) {
@@ -322,15 +366,14 @@ impl fpm::Config {
         }
         *files = files
             .iter_mut()
-            .enumerate()
             .filter_map(|(k, v)| {
-                if !remove_files.contains(&k) {
-                    Some(v.to_owned())
+                if !uptodate_files.contains(k) {
+                    Some((k.to_owned(), v.to_owned()))
                 } else {
                     None
                 }
             })
-            .collect_vec();
-        Ok(())
+            .collect();
+        Ok((already_added_files, already_removed_files))
     }
 }
